mcp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
azure-ai-inference==1.0.0b9
azure-core==1.35.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
fastapi==0.118.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
idna==3.10
isodate==0.7.2
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mcp==1.16.0
mdurl==0.2.0
multidict==6.7.0
openai==2.2.0
propcache==0.4.1
pydantic==2.12.0
pydantic_core==2.41.1
pydantic-settings==2.11.0
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tqdm==4.67.1
typer==0.19.2
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.22.0
